Refactor types package into sub packages
