+++ src/main/java/com/amazon/opendistroforelasticsearch/sql/antlr/semantic/types/TypeExpression.java
@@ -15,6 +15,7 @@
 
 package com.amazon.opendistroforelasticsearch.sql.antlr.semantic.types;
 
+import com.amazon.opendistroforelasticsearch.sql.antlr.semantic.types.special.Generic;
 import com.amazon.opendistroforelasticsearch.sql.utils.StringUtils;
 
 import java.util.Arrays;
@@ -22,8 +23,8 @@ import java.util.List;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import static com.amazon.opendistroforelasticsearch.sql.antlr.semantic.types.BaseType.TYPE_ERROR;
-import static com.amazon.opendistroforelasticsearch.sql.antlr.semantic.types.BaseType.UNKNOWN;
+import static com.amazon.opendistroforelasticsearch.sql.antlr.semantic.types.base.BaseType.TYPE_ERROR;
+import static com.amazon.opendistroforelasticsearch.sql.antlr.semantic.types.base.BaseType.UNKNOWN;
 
 /**
  * Type expression representing specification(s) of constructor such as function, operator etc.
