+++ src/main/java/com/amazon/opendistroforelasticsearch/sql/antlr/semantic/types/special/Product.java
@@ -13,7 +13,9 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.antlr.semantic.types;
+package com.amazon.opendistroforelasticsearch.sql.antlr.semantic.types.special;
+
+import com.amazon.opendistroforelasticsearch.sql.antlr.semantic.types.Type;
 
 import java.util.Collections;
 import java.util.List;
